mv 2 test cases which can not parallel to separate file
